add redisMessageListenerContainer, update DataSeeder

+++ Backend/src/main/java/com/ticketrush/backend/repository/SeatTypeRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface SeatTypeRepository extends JpaRepository<SeatType, Double> {
+public interface SeatTypeRepository extends JpaRepository<SeatType, Integer> {
     Optional<SeatType> findByName(String name);
 
     Boolean existsByName(String name);
